# Remove full jacobian from stepper